pyproject.toml
[project]
name = "vaes"
version = "0.1.0"
description = "Variational AutoEncoders on MNIST"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch",
    "torchvision",
    "hydra-core",
    "omegaconf",
    "mlflow",
    "python-dotenv",
    "tqdm",
    "matplotlib",
    "numpy",
]

[project.scripts]
train = "vaes.cli.train:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vaes"]
